third_party/WebKit/LayoutTests/fast/css/media-query-in-html-import.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="import" href="data:text/html,<style>@media (max-width: 0px){div{background:red}}</style>">
<div></div>
<script>
    test(() => {
        assert_equals(getComputedStyle(document.querySelector("div")).backgroundColor, "rgba(0, 0, 0, 0)");
    }, "(max-width: 0px) media query in import document should not evaluate to true.");
</script>
